Edited bases and done examples
    - New 5 examples with UVMEnv 2.0
    - Edited doc about user interface. Now is ready the new user interface.
    - Edited bases for well coding.

+++ src/implements/uvmenv_pseudo/objects/SequenceItem.cpp
@@ -87,7 +87,7 @@ void SequenceItem::copyBaseFile(){
     }, PATH_SEP));
     string topModule = config["dut_design"]["top_module"];
 
-    long randMax = 0;
+    long long randMax = 0;
     for(const auto& [module, signalProps] : dutSignals) {
         if(module == topModule){
             for (const auto& signal : signalProps) {
